initial 7/13/2019

+++ app/src/main/java/com/ucsdextandroid2/petfinder/DataSource.kt
@@ -91,7 +91,7 @@ object DataSource {
     }
 
 
-    fun findAnimals(lat: Long?, lng: Long?, page: Int = 1, limit: Int = 20, resultListener: (ApiResult<AnimalsResponse>) -> Unit) {
+    fun findAnimals(lat: Double?, lng: Double?, page: Int = 1, limit: Int = 20, resultListener: (ApiResult<AnimalsResponse>) -> Unit) {
         createApi().getAnimals(
             location = if(lat != null && lng != null) "$lat,$lng" else null,
             page = page,
